+++ saleor/static/dashboard-next/discounts/components/DiscountCountrySelectDialog/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DiscountCountrySelectDialog';
+export * from './DiscountCountrySelectDialog';