+++ saleor/static/dashboard-next/discounts/components/SaleSummary/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./SaleSummary";
+export * from "./SaleSummary";