+++ saleor/static/dashboard-next/types/globalTypes.ts
@@ -15,6 +15,11 @@ export enum AuthorizationKeyType {
   GOOGLE_OAUTH2 = "GOOGLE_OAUTH2",
 }
 
+export enum DiscountValueTypeEnum {
+  FIXED = "FIXED",
+  PERCENTAGE = "PERCENTAGE",
+}
+
 export enum FulfillmentStatus {
   CANCELED = "CANCELED",
   FULFILLED = "FULFILLED",
@@ -86,6 +91,11 @@ export enum PermissionEnum {
   MANAGE_USERS = "MANAGE_USERS",
 }
 
+export enum SaleType {
+  FIXED = "FIXED",
+  PERCENTAGE = "PERCENTAGE",
+}
+
 export enum StockAvailability {
   IN_STOCK = "IN_STOCK",
   OUT_OF_STOCK = "OUT_OF_STOCK",
@@ -117,6 +127,27 @@ export enum TaxRateType {
   STANDARD = "STANDARD",
   WATER = "WATER",
   WINE = "WINE",
+}
+
+export enum VoucherDiscountValueType {
+  FIXED = "FIXED",
+  PERCENTAGE = "PERCENTAGE",
+}
+
+export enum VoucherType {
+  CATEGORY = "CATEGORY",
+  COLLECTION = "COLLECTION",
+  PRODUCT = "PRODUCT",
+  SHIPPING = "SHIPPING",
+  VALUE = "VALUE",
+}
+
+export enum VoucherTypeEnum {
+  CATEGORY = "CATEGORY",
+  COLLECTION = "COLLECTION",
+  PRODUCT = "PRODUCT",
+  SHIPPING = "SHIPPING",
+  VALUE = "VALUE",
 }
 
 export enum WeightUnitsEnum {
@@ -164,6 +195,12 @@ export interface AttributeValueInput {
 export interface AuthorizationKeyInput {
   key: string;
   password: string;
+}
+
+export interface CatalogueInput {
+  products?: (string | null)[] | null;
+  categories?: (string | null)[] | null;
+  collections?: (string | null)[] | null;
 }
 
 export interface CategoryInput {
@@ -281,6 +318,17 @@ export interface ProductVariantInput {
   weight?: any | null;
 }
 
+export interface SaleInput {
+  name?: string | null;
+  type?: DiscountValueTypeEnum | null;
+  value?: any | null;
+  products?: (string | null)[] | null;
+  categories?: (string | null)[] | null;
+  collections?: (string | null)[] | null;
+  startDate?: any | null;
+  endDate?: any | null;
+}
+
 export interface SeoInput {
   title?: string | null;
   description?: string | null;
@@ -329,6 +377,21 @@ export interface UserCreateInput {
   isActive?: boolean | null;
   note?: string | null;
   sendPasswordEmail?: boolean | null;
+}
+
+export interface VoucherInput {
+  type?: VoucherTypeEnum | null;
+  name?: string | null;
+  code?: string | null;
+  startDate?: any | null;
+  endDate?: any | null;
+  discountValueType?: DiscountValueTypeEnum | null;
+  discountValue?: any | null;
+  products?: (string | null)[] | null;
+  collections?: (string | null)[] | null;
+  categories?: (string | null)[] | null;
+  minAmountSpent?: any | null;
+  countries?: (string | null)[] | null;
 }
 
 //==============================================================